no need to load wininet.dll on runtime as it is available on all supported platforms by default
git-svn-id: file:///.data/svn_convert/svn@6879 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ branches/branch-QString/sim/plugins.cpp
@@ -312,7 +312,7 @@ pluginInfo *PluginManagerPrivate::getInfo(const QString &name)
 
 void PluginManagerPrivate::release_all(Plugin *to)
 {
-    for (size_t n = plugins.size() - 1; n >= 0; n--){
+    for (int n = plugins.size() - 1; n >= 0; n--){
         pluginInfo &info = plugins[n];
         if (to && (info.plugin == to))
             break;
